requirements.txt
click==7.1.2
python-dotenv==1.0.1
python-decouple==3.8
